add redirect to Login page if haven't token

+++ client/client-react/src/Admin/UserManager/userManager.tsx
@@ -84,7 +84,8 @@ class UserManager extends React.Component<any, State> {
   }
 
   public deleteUserFromDB(user: User) {
-    if ((this.userInfoService.getCurrentUser() as User).id !== user._id ) {
+
+    if (this.userInfoService.getCurrentUser() && (this.userInfoService.getCurrentUser() as User).id !== user._id ) {
       this.userService.delete(user._id)
       .then(() => {
         alert('successfuly deleted');
